+++ src/main-curriculum/endpoints/axios-instance.ts
@@ -4,4 +4,7 @@ const BASE_URL = import.meta.env.VITE_BASE_URL;
 
 export const instance = axios.create({
   baseURL: BASE_URL,
+  headers: {
+    "Content-Type": "application/json",
+  },
 });