+++ src/main-curriculum/endpoints/index.ts
@@ -3,3 +3,4 @@ export * from "./personal-info";
 export * from "./about-me";
 export * from "./sections";
 export * from "./soft-skills";
+export * from "./curriculum-pdf";