+++ src/main-curriculum/helpers/icons.tsx
@@ -4,6 +4,7 @@ import { IoSchoolSharp } from "react-icons/io5";
 import { FaBriefcase } from "react-icons/fa";
 import { MdVolunteerActivism } from "react-icons/md";
 import { GiSkills } from "react-icons/gi";
+import { FaFilePdf } from "react-icons/fa6";
 
 export const icons: Record<string, IconType> = {
   personBadge: BsFillPersonBadgeFill,
@@ -11,6 +12,7 @@ export const icons: Record<string, IconType> = {
   workExperience: FaBriefcase,
   volunteering: MdVolunteerActivism,
   softSkills: GiSkills,
+  downloadCV: FaFilePdf,
 };
 
 export const getIconFromIconKey = (iconKey: string) => {